frontend/src/app/services/currentStockService/stock-shared-service.service.spec.ts
import { TestBed } from '@angular/core/testing';

import { StockSharedServiceService } from './stock-shared-service.service';

/**
 * Test Suite: StockSharedServiceService
 * 
 * Filename: stock-shared-service.service.spec.ts
 * Author: [Cole Baxendale]
 * Contact: [thecodercole@gmail.com]
 * Created on: [February 2024]
 * Version: 1.0
 * 
 * Description: 
 * This file contains the test suite for the StockSharedServiceService. It aims to ensure the integrity and functionality
 * of the StockSharedServiceService, verifying that the service is properly instantiated and behaves as expected within
 * the application. This suite focuses on testing the foundational capabilities of the service, such as its ability to
 * be created and to manage stock details data effectively.
 * 
 * Test Strategy:
 * - Initialization test to verify that the service can be successfully created and injected.
 * - (Further tests could be designed to assess the functionality of methods like updateCurrentStockDetails and
 *    clearCurrentStockDetails, ensuring they correctly manipulate the BehaviorSubject and notify subscribers.)
 */

describe('StockSharedServiceService', () => {
  let service: StockSharedServiceService;

  beforeEach(() => {
    TestBed.configureTestingModule({}); // Setup TestBed with the necessary environment for the service
    service = TestBed.inject(StockSharedServiceService); // Inject the service to be tested
  });

  it('should be created', () => {
    // Test to ensure the service instance is successfully created
    expect(service).toBeTruthy();
  });

  // Additional tests can be added here to further verify the service's methods and their effects
});
